Extract widgets from onboarding feature
- Extract OnboardingTermsScreen, OnboardingWelcomeScreen, and OnboardingErrorScreen widgets
- Create widgets directory structure for onboarding feature
- Update onboarding_screen.dart to use extracted widgets
- Clean up unused imports in extracted widgets
- Maintain widget extraction pattern consistency across features
