+++ lib/features/chat/widgets/chat_app_bar_title.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../../core/constants/app_constants.dart';
 import '../../settings/bloc/agent_bloc.dart';
 import '../../settings/bloc/agent_states.dart';
