+++ lib/features/file_management/presentation/file_management_screen.dart
@@ -3,9 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import '../bloc/file_bloc.dart';
 import '../bloc/file_events.dart';
 import '../bloc/file_states.dart';
-import '../models/file_model.dart';
 import '../../../shared/widgets/loading_indicator.dart';
-import '../../../core/constants/app_constants.dart';
 import '../widgets/file_error_message.dart';
 import '../widgets/file_empty_state.dart';
 import '../widgets/file_list.dart';
